moved matchmedia mock function to mocks
