+++ src/App.test.tsx
@@ -1,4 +1,4 @@
-import './matchMediaMock';
+import 'mocks/matchMediaMock';
 import { MemoryRouter } from 'react-router-dom';
 import { MockedProvider } from '@apollo/client/testing';
 import { waitFor, render } from '@testing-library/react';
