+++ src/components/UI/Form/EmojiInput/EmojiInput.test.tsx
@@ -1,4 +1,4 @@
-import '../../../../matchMediaMock';
+import 'mocks/matchMediaMock';
 import { render } from '@testing-library/react';
 import { EmojiInput } from './EmojiInput';
 import userEvent from '@testing-library/user-event';
