+++ src/containers/Chat/ChatInterface/ChatInterface.test.tsx
@@ -1,4 +1,4 @@
-import '../../../matchMediaMock';
+import 'mocks/matchMediaMock';
 import { MemoryRouter } from 'react-router-dom';
 import { cleanup, render } from '@testing-library/react';
 import { ApolloClient, ApolloProvider, InMemoryCache } from '@apollo/client';
