+++ src/containers/Chat/ChatMessages/ChatMessages.test.tsx
@@ -1,4 +1,4 @@
-import '../../../matchMediaMock';
+import 'mocks/matchMediaMock';
 import { render, screen, act } from '@testing-library/react';
 import { ApolloClient, ApolloProvider, InMemoryCache } from '@apollo/client';
 import { fireEvent, waitFor } from '@testing-library/dom';
