+++ src/containers/InteractiveMessage/InteractiveMessage.test.tsx
@@ -1,4 +1,4 @@
-import '../../matchMediaMock';
+import 'mocks/matchMediaMock';
 import { render, screen, waitFor, fireEvent } from '@testing-library/react';
 import { MockedProvider } from '@apollo/client/testing';
 import axios from 'axios';
