+++ src/routes/AuthenticatedRoute/AuthenticatedRoute.test.tsx
@@ -1,4 +1,4 @@
-import '../../matchMediaMock';
+import 'mocks/matchMediaMock';
 import { Suspense } from 'react';
 import { render, waitFor } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
